datasets and quality added

+++ Report2/common/input.py
@@ -2,5 +2,5 @@ import numpy as np
 import csv
 
 def parse_txt(path) -> np.ndarray:
-    data = np.loadtxt(path, dtype=int)
+    data = np.loadtxt(path, dtype = np.float64)
     return data
